+++ app/src/main/java/org/eyeseetea/malariacare/database/utils/monitor/allassessments/EntrySentSurveysChartBase.java
@@ -17,7 +17,7 @@
  *  along with Foobar.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.eyeseetea.malariacare.database.utils.monitor.allassessment;
+package org.eyeseetea.malariacare.database.utils.monitor.allassessments;
 
 import java.text.SimpleDateFormat;
 import java.util.Date;
