+++ app/src/main/java/org/eyeseetea/malariacare/database/utils/monitor/allassessments/EntrySentSurveysChartByOrgUnit.java
@@ -17,26 +17,25 @@
  *  along with Foobar.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.eyeseetea.malariacare.database.utils.monitor.allassessment;
+package org.eyeseetea.malariacare.database.utils.monitor.allassessments;
 
-import org.eyeseetea.malariacare.database.model.Program;
+import org.eyeseetea.malariacare.database.model.OrgUnit;
 
 import java.util.Date;
 
 /**
  * Created by idelcano on 23/08/2016.
  */
-public class EntrySentSurveysChartByProgram extends  EntrySentSurveysChartBase {
-    public static final String JAVASCRIPT_ADD_DATA = "javascript:setProgramData([%d, %d, '%s', '%s','%s'])";
-
+public class EntrySentSurveysChartByOrgUnit  extends  EntrySentSurveysChartBase{
+    public static final String JAVASCRIPT_ADD_DATA = "javascript:setOrgUnitData([%d, %d, '%s', '%s','%s'])";
 
     /**
      * The date whose month represents this point (day info is discarded)
      */
-    Program program;
-    public EntrySentSurveysChartByProgram(int expected, Date date, Program program) {
+    OrgUnit orgUnit;
+    public EntrySentSurveysChartByOrgUnit(int expected, Date date,OrgUnit orgUnit) {
         super(expected, date);
-        this.program=program;
+        this.orgUnit=orgUnit;
     }
 
     /**
@@ -44,6 +43,6 @@ public class EntrySentSurveysChartByProgram extends  EntrySentSurveysChartBase {
      * @return
      */
     public String getEntryAsJS(){
-        return String.format(JAVASCRIPT_ADD_DATA,sent,expected,program.getName(),program.getUid(),getDateAsString());
+        return String.format(JAVASCRIPT_ADD_DATA,sent,expected,orgUnit.getName(),orgUnit.getUid(),getDateAsString());
     }
 }
