+++ app/src/main/java/org/eyeseetea/malariacare/fragments/MonitorFragment.java
@@ -44,12 +44,12 @@ import org.eyeseetea.malariacare.database.utils.Session;
 import org.eyeseetea.malariacare.database.utils.monitor.MonitorMessagesBuilder;
 import org.eyeseetea.malariacare.database.utils.monitor.pie.PieBuilderBase;
 import org.eyeseetea.malariacare.database.utils.services.BaseServiceBundle;
-import org.eyeseetea.malariacare.database.utils.monitor.allassessment.SentSurveysBuilderByOrgUnit;
-import org.eyeseetea.malariacare.database.utils.monitor.allassessment.SentSurveysBuilderByProgram;
+import org.eyeseetea.malariacare.database.utils.monitor.allassessments.SentSurveysBuilderByOrgUnit;
+import org.eyeseetea.malariacare.database.utils.monitor.allassessments.SentSurveysBuilderByProgram;
 import org.eyeseetea.malariacare.database.utils.monitor.facility.FacilityTableBuilderBase;
 import org.eyeseetea.malariacare.database.utils.monitor.facility.FacilityTableBuilderByOrgUnit;
 import org.eyeseetea.malariacare.database.utils.monitor.facility.FacilityTableBuilderByProgram;
-import org.eyeseetea.malariacare.database.utils.monitor.allassessment.SentSurveysBuilderBase;
+import org.eyeseetea.malariacare.database.utils.monitor.allassessments.SentSurveysBuilderBase;
 import org.eyeseetea.malariacare.database.utils.monitor.pie.PieBuilderByOrgUnit;
 import org.eyeseetea.malariacare.database.utils.monitor.pie.PieBuilderByProgram;
 import org.eyeseetea.malariacare.layout.adapters.dashboard.IDashboardAdapter;
